requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
numpy>=1.24.0

# FastAPI backend dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0

# OpenAI integration
openai>=1.3.0

# Audio processing
st-audiorec>=0.1.4
pydub>=0.25.1

# Enhanced Streamlit components
streamlit-js-eval>=0.1.6
streamlit-extras>=0.3.0
